Send remaining 96% of the data to misp as well
This is rather sizable change read on.

Previously, we would sent only first 20 items of the 5000 batch to misp, this
was due to an error of indentation.

Please note that when code leaves the `with executor` block as follows

    with concurrent.futures.ThreadPoolExecutor() as executor:
        pass

the executor is closed for new additions to the queue. So, we have previously
only sent 20 items from the 5000 long batch to the executor and we silently
failed to add remaining 4890.